requirements.txt
django == 2.0.4
djangorestframework == 3.8.2
mysqlclient == 1.3.12
PyMySQL == 0.8.0
django-cors-headers == 2.2.0
gunicorn==19.7.1
whitenoise==3.3.1